Некоторые правки в коде
